test: split oversized dump-depth test file, close dry_run_compile_db_matched coverage gap
- tests/test_cli_dump_helpers_coverage.py grew past the 2000-line
  AI-readiness hard cap over this session's rounds of additions; splits
  the depth/provenance test block (evidence_depth_label,
  check_requested_depth_satisfied, _gated_source_label,
  fold_dump_provenance_into_json, _l4_source_abi_was_attempted,
  _dump_will_attempt_hybrid_l4_extraction) into a new sibling file,
  tests/test_dump_depth_provenance.py.
- Adds direct unit tests for dry_run_compile_db_matched's branches that
  Codecov flagged as uncovered: the union-fallback (no headers) path,
  and the malformed/missing-compile-db exception path that must fold
  into False rather than propagate.
